+++ packages/sphinx-test-reports/tests/doc_test/basic_doc/conf.py
@@ -72,7 +72,7 @@ language = "en"
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
-if Version(sphinx_needs.__version__) >= Version("8.0.0"):
+if Version(sphinx_needs.__version__) >= Version("7.0.0"):
     needs_fields = {"more_info": {"nullable": True}}
 else:
     needs_extra_options = ["more_info"]