+++ packages/sphinx-test-reports/tests/doc_test/testsuites_doc/conf.py
@@ -37,7 +37,7 @@ extensions = ["sphinx_needs", "sphinxcontrib.test_reports"]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
-if Version(sphinx_needs.__version__) >= Version("8.0.0"):
+if Version(sphinx_needs.__version__) >= Version("7.0.0"):
     needs_fields = {"more_info": {"nullable": True}}
 else:
     needs_extra_options = ["more_info"]